Fix fake movements from Wanway S20 tracker
Sometimes, the tracker sends false movements when the vehicle is stopped. This error usually occurs when the vehicle is stopped on a steep slope. To fix this, we ensure that even if the speed value is indicated as positive, if the position coordinates do not change within 30 milliseconds, the speed value is forced to zero.