Changed the use of the geocoder.reuseDistance parameter
The "geocoder.reuseDistance" parameter used in "/opt/traccar/conf/traccar.xml" has been changed so that instead of reusing the address obtained from the previous request depending on the distance, it is only required to reuse it if the vehicle's position coordinates haven't changed. This is intended to reduce the number of requests to the geocoding API, while also ensuring the address is always updated when the position changes.